Adjusted all classes and used rxbus to button click
