+++ app/src/main/java/com/example/anilreddy/rxjava/model/Events.java
@@ -1,0 +1,16 @@
+package com.example.anilreddy.rxjava.model;
+
+public class Events {
+
+    public Events() {
+
+    }
+
+    public static class TapEvent {
+
+    }
+
+    public static class AutoEvent {
+
+    }
+}
