--- app/src/main/java/com/example/anilreddy/rxjava/operations/startTimerActivity.java
@@ -1,78 +1,0 @@
-package com.example.anilreddy.rxjava.operations;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.widget.Button;
-import android.widget.TextView;
-
-import com.example.anilreddy.rxjava.R;
-import com.example.anilreddy.rxjava.utils.AppConstant;
-
-import java.util.concurrent.TimeUnit;
-
-import io.reactivex.Observable;
-import io.reactivex.Observer;
-import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.disposables.Disposable;
-import io.reactivex.schedulers.Schedulers;
-
-public class startTimerActivity extends AppCompatActivity {
-
-    private final static String TAG = startTimerActivity.class.getSimpleName();
-    Button btn;
-    TextView textView;
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_example_sample);
-
-        btn = findViewById(R.id.btn);
-        textView = findViewById(R.id.textView);
-
-        btn.setOnClickListener((view) -> doSomeWork());
-    }
-
-    private void doSomeWork() {
-        getObservable().subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(getObserver());
-
-    }
-
-    private Observable<Long> getObservable() {
-        return Observable.timer(2, TimeUnit.SECONDS);
-    }
-
-    private Observer<Long> getObserver() {
-        return new Observer<Long>() {
-            @Override
-            public void onSubscribe(Disposable d) {
-                Log.d(TAG, " onSubscribe : " + d.isDisposed());
-            }
-
-            @Override
-            public void onNext(Long value) {
-                textView.append(" onNext : value : " + value);
-                textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onNext : value : " + value);
-            }
-
-            @Override
-            public void onError(Throwable e) {
-                textView.append(" onError : " + e.getMessage());
-                textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onError : " + e.getMessage());
-            }
-
-            @Override
-            public void onComplete() {
-                textView.append(" onComplete : ");
-                textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onComplete : ");
-            }
-        };
-    }
-}
