--- app/src/main/java/com/example/anilreddy/rxjava/rxbus/RxBus.java
@@ -1,4 +1,0 @@
-package com.example.anilreddy.rxjava.rxbus;
-
-public class RxBus {
-}
