--- app/src/main/java/com/example/anilreddy/rxjava/rxbus/RxBusActivity.java
@@ -1,6 +1,0 @@
-package com.example.anilreddy.rxjava.rxbus;
-
-import android.support.v7.app.AppCompatActivity;
-
-public class RxBusActivity extends AppCompatActivity {
-}
