+++ app/src/main/java/com/example/anilreddy/rxjava/ui/operations/DisposableExampleActivity.java
@@ -1,4 +1,4 @@
-package com.example.anilreddy.rxjava.operations;
+package com.example.anilreddy.rxjava.ui.operations;
 
 import android.os.Bundle;
 import android.os.SystemClock;
