+++ app/src/main/java/com/example/anilreddy/rxjava/ui/operations/MapExampleActivity.java
@@ -1,4 +1,4 @@
-package com.example.anilreddy.rxjava.operations;
+package com.example.anilreddy.rxjava.ui.operations;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -7,6 +7,7 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.example.anilreddy.rxjava.R;
+import com.example.anilreddy.rxjava.model.ApiUser;
 import com.example.anilreddy.rxjava.model.User;
 import com.example.anilreddy.rxjava.utils.AppConstant;
 import com.example.anilreddy.rxjava.utils.Utils;
@@ -17,18 +18,20 @@ import io.reactivex.Observable;
 import io.reactivex.Observer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
+import io.reactivex.functions.Function;
 import io.reactivex.schedulers.Schedulers;
 
-public class ZipExampleActivity extends AppCompatActivity {
+public class MapExampleActivity extends AppCompatActivity {
 
-    public static final String TAG = ZipExampleActivity.class.getSimpleName();
+    private static final String TAG = SimpleExampleActivity.class.getName();
     Button btn;
     TextView textView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_zip_example);
+        setContentView(R.layout.activity_map_example);
+
         btn = findViewById(R.id.btn);
         textView = findViewById(R.id.textView);
 
@@ -36,28 +39,17 @@ public class ZipExampleActivity extends AppCompatActivity {
     }
 
     private void doSomeWork() {
-        Observable.zip(getCricketFansObservable(), getFootballFansObservable(),
-                //                (cricketFan, footballFan) ->
-                //                    Utils.filterUserWhoLovesBoth(cricketFan, footballFan)
-                Utils::filterUserWhoLovesBoth)
-                .subscribeOn(Schedulers.io())
+        getObservable().subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(getObserver());
+                .map((Function<List<ApiUser>, List<User>>) apiUser
+                        -> Utils.convertApiUserListToUserList(apiUser)).subscribe(getObserver());
+
     }
 
-    public Observable<List<User>> getCricketFansObservable() {
+    public Observable<List<ApiUser>> getObservable() {
         return Observable.create((e) -> {
             if (!e.isDisposed()) {
-                e.onNext(Utils.getUserListWhoLovesCricket());
-                e.onComplete();
-            }
-        });
-    }
-
-    public Observable<List<User>> getFootballFansObservable() {
-        return Observable.create((e) -> {
-            if (!e.isDisposed()) {
-                e.onNext(Utils.getUserListWhoLovesFootball());
+                e.onNext(Utils.getApiUserList());
                 e.onComplete();
             }
         });
@@ -72,20 +64,20 @@ public class ZipExampleActivity extends AppCompatActivity {
 
             @Override
             public void onNext(List<User> users) {
-                textView.append(" onNext:");
+                textView.append(" onNext");
                 textView.append(AppConstant.LINE_SEPARATOR);
                 for (User user : users) {
-                    textView.append(" firstName: " + user.firstName);
+                    textView.append("firstName: " + user.firstName);
                     textView.append(AppConstant.LINE_SEPARATOR);
                 }
-                Log.d(TAG, " onNext : " + users.size());
+                Log.d(TAG, "onNext : " + users.size());
             }
 
             @Override
             public void onError(Throwable e) {
-                textView.append(" onError : " + e.getMessage());
+                textView.append(" onError" + e.getMessage());
                 textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onError : " + e.getMessage());
+                Log.d(TAG, " onError: " + e.getMessage());
             }
 
             @Override
@@ -96,4 +88,5 @@ public class ZipExampleActivity extends AppCompatActivity {
             }
         };
     }
+
 }
