+++ app/src/main/java/com/example/anilreddy/rxjava/ui/operations/SingleObserverExampleActivity.java
@@ -1,4 +1,4 @@
-package com.example.anilreddy.rxjava.operations;
+package com.example.anilreddy.rxjava.ui.operations;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -10,15 +10,13 @@ import android.widget.TextView;
 import com.example.anilreddy.rxjava.R;
 import com.example.anilreddy.rxjava.utils.AppConstant;
 
-import io.reactivex.Observable;
-import io.reactivex.Observer;
-import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.Single;
+import io.reactivex.SingleObserver;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.schedulers.Schedulers;
 
-public class TakeExampleActivity extends AppCompatActivity {
+public class SingleObserverExampleActivity extends AppCompatActivity {
 
-    private static final String TAG = TakeExampleActivity.class.getSimpleName();
+    public static final String TAG = SingleObserverExampleActivity.class.getSimpleName();
     Button btn;
     TextView textView;
 
@@ -30,32 +28,26 @@ public class TakeExampleActivity extends AppCompatActivity {
         btn = findViewById(R.id.btn);
         textView = findViewById(R.id.textView);
 
-        btn.setOnClickListener((view) -> doSomeWork());
+        btn.setOnClickListener((v) -> doSomeWork());
     }
 
     private void doSomeWork() {
-        getObservable().subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .take(3)
-                .subscribe(getObserver());
+        Single.just("Amit").subscribe(getSingleObserver());
+
     }
 
-    private Observable<Integer> getObservable() {
-        return Observable.just(1, 2, 3, 4, 5);
-    }
-
-    private Observer<Integer> getObserver() {
-        return new Observer<Integer>() {
+    private SingleObserver<String> getSingleObserver() {
+        return new SingleObserver<String>() {
             @Override
             public void onSubscribe(Disposable d) {
                 Log.d(TAG, " onSubscribe : " + d.isDisposed());
             }
 
             @Override
-            public void onNext(Integer value) {
-                textView.append(" onNext : value : " + value);
+            public void onSuccess(String s) {
+                textView.append(" onNext : value : " + s);
                 textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onNext value : " + value);
+                Log.d(TAG, " onNext value : " + s);
             }
 
             @Override
@@ -63,13 +55,6 @@ public class TakeExampleActivity extends AppCompatActivity {
                 textView.append(" onError : " + e.getMessage());
                 textView.append(AppConstant.LINE_SEPARATOR);
                 Log.d(TAG, " onError : " + e.getMessage());
-            }
-
-            @Override
-            public void onComplete() {
-                textView.append(" onComplete");
-                textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onComplete");
             }
         };
     }
