+++ app/src/main/java/com/example/anilreddy/rxjava/ui/operations/TakeExampleActivity.java
@@ -1,36 +1,31 @@
-package com.example.anilreddy.rxjava.operations;
+package com.example.anilreddy.rxjava.ui.operations;
 
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.Button;
 import android.widget.TextView;
 
 import com.example.anilreddy.rxjava.R;
-import com.example.anilreddy.rxjava.model.ApiUser;
-import com.example.anilreddy.rxjava.model.User;
 import com.example.anilreddy.rxjava.utils.AppConstant;
-import com.example.anilreddy.rxjava.utils.Utils;
-
-import java.util.List;
 
 import io.reactivex.Observable;
 import io.reactivex.Observer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Function;
 import io.reactivex.schedulers.Schedulers;
 
-public class MapExampleActivity extends AppCompatActivity {
+public class TakeExampleActivity extends AppCompatActivity {
 
-    private static final String TAG = SimpleExampleActivity.class.getName();
+    private static final String TAG = TakeExampleActivity.class.getSimpleName();
     Button btn;
     TextView textView;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_map_example);
+        setContentView(R.layout.activity_example_sample);
 
         btn = findViewById(R.id.btn);
         textView = findViewById(R.id.textView);
@@ -41,43 +36,33 @@ public class MapExampleActivity extends AppCompatActivity {
     private void doSomeWork() {
         getObservable().subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .map((Function<List<ApiUser>, List<User>>) apiUser
-                        -> Utils.convertApiUserListToUserList(apiUser)).subscribe(getObserver());
-
+                .take(3)
+                .subscribe(getObserver());
     }
 
-    public Observable<List<ApiUser>> getObservable() {
-        return Observable.create((e) -> {
-            if (!e.isDisposed()) {
-                e.onNext(Utils.getApiUserList());
-                e.onComplete();
-            }
-        });
+    private Observable<Integer> getObservable() {
+        return Observable.just(1, 2, 3, 4, 5);
     }
 
-    public Observer<List<User>> getObserver() {
-        return new Observer<List<User>>() {
+    private Observer<Integer> getObserver() {
+        return new Observer<Integer>() {
             @Override
             public void onSubscribe(Disposable d) {
                 Log.d(TAG, " onSubscribe : " + d.isDisposed());
             }
 
             @Override
-            public void onNext(List<User> users) {
-                textView.append(" onNext");
+            public void onNext(Integer value) {
+                textView.append(" onNext : value : " + value);
                 textView.append(AppConstant.LINE_SEPARATOR);
-                for (User user : users) {
-                    textView.append("firstName: " + user.firstName);
-                    textView.append(AppConstant.LINE_SEPARATOR);
-                }
-                Log.d(TAG, "onNext : " + users.size());
+                Log.d(TAG, " onNext value : " + value);
             }
 
             @Override
             public void onError(Throwable e) {
-                textView.append(" onError" + e.getMessage());
+                textView.append(" onError : " + e.getMessage());
                 textView.append(AppConstant.LINE_SEPARATOR);
-                Log.d(TAG, " onError: " + e.getMessage());
+                Log.d(TAG, " onError : " + e.getMessage());
             }
 
             @Override
@@ -88,5 +73,4 @@ public class MapExampleActivity extends AppCompatActivity {
             }
         };
     }
-
 }
