+++ app/src/main/java/com/example/anilreddy/rxjava/ui/rxbus/RxBusActivity.java
@@ -1,0 +1,61 @@
+package com.example.anilreddy.rxjava.ui.rxbus;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+import android.widget.TextView;
+
+import com.example.anilreddy.rxjava.MyApplication;
+import com.example.anilreddy.rxjava.R;
+import com.example.anilreddy.rxjava.model.Events;
+
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.functions.Consumer;
+import io.reactivex.schedulers.Schedulers;
+
+public class RxBusActivity extends AppCompatActivity {
+
+    public static final String TAG = RxBusActivity.class.getSimpleName();
+    TextView textView;
+    Button button;
+    private final CompositeDisposable disposables = new CompositeDisposable();
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        disposables.clear();
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_rxbus);
+        textView = findViewById(R.id.textView);
+        button = findViewById(R.id.button);
+
+        disposables.add(((MyApplication) getApplication())
+                .bus()
+                .toObservable()
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Consumer<Object>() {
+                    @Override
+                    public void accept(Object o) throws Exception {
+                        if (o instanceof Events.AutoEvent) {
+                            textView.setText("Auto Event Received");
+                        } else if (o instanceof Events.TapEvent) {
+                            textView.setText("Tap Event Received");
+                        }
+                    }
+                }));
+
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                ((MyApplication) getApplication()).bus().send(new Events.TapEvent());
+            }
+        });
+    }
+}
