+++ app/src/main/res/layout/activity_operators.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".OperatorsActivity">
+    tools:context=".ui.OperatorsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
