+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".search.SearchActivity">
+    tools:context=".ui.search.SearchActivity">
 
     <SearchView
         android:id="@+id/searchView"
