+++ app/src/main/res/layout/activity_zip_example.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".operations.ZipExampleActivity">
+    tools:context=".ui.operations.ZipExampleActivity">
 
     <Button
         android:id="@+id/btn"
